[chore] Home Layout func Extension으로 분리
