+++ B-umiOS/B-umiOS/Home/HomeViewController.swift
@@ -15,19 +15,19 @@ import UIKit
 class HomeViewController: UIViewController {
     // MARK: - UIComponenets
     
-    private let backgroundView = AnimationView().then {
+    let backgroundView = AnimationView().then {
         $0.animation = Animation.named("home_ios")
         $0.loopMode = .playOnce
     }
 
-    private let dateLabel = UILabel().then {
+    let dateLabel = UILabel().then {
         $0.text = Date().dateToString(date: Date())
         $0.font = UIFont.nanumSquareFont(type: .extraBold, size: 20)
         $0.adjustsFontForContentSizeCategory = true
         $0.textColor = .white
     }
     
-    private let guideLabel = UILabel().then {
+    let guideLabel = UILabel().then {
         $0.text = "휴지통을 클릭해\n스트레스를 비워보세요"
         $0.lineSpacing(spacing: 13)
         $0.numberOfLines = 2
@@ -37,16 +37,16 @@ class HomeViewController: UIViewController {
         $0.textColor = UIColor.white
     }
     
-    private let arrowImage = UIImageView().then {
+    let arrowImage = UIImageView().then {
         $0.image = UIImage(named: "homeAreaTrashbin")
     }
     
-    private lazy var trashBinButton = UIButton().then {
+    lazy var trashBinButton = UIButton().then {
         $0.addTarget(self, action: #selector(didTapTrashBinButton(_:)), for: .touchUpInside)
         $0.backgroundColor = .clear
     }
     
-    private lazy var paper1Button = UIButton().then {
+    lazy var paper1Button = UIButton().then {
         $0.addTarget(self, action: #selector(didTapPaperButton(_:)), for: .touchUpInside)
         $0.backgroundColor = .paper1
         $0.tag = 3
@@ -54,7 +54,7 @@ class HomeViewController: UIViewController {
         $0.isHidden = true
     }
     
-    private lazy var paper2Button = UIButton().then {
+    lazy var paper2Button = UIButton().then {
         $0.addTarget(self, action: #selector(didTapPaperButton(_:)), for: .touchUpInside)
         $0.backgroundColor = .paper2
         $0.tag = 2
@@ -62,7 +62,7 @@ class HomeViewController: UIViewController {
         $0.isHidden = true
     }
     
-    private lazy var paper3Button = UIButton().then {
+    lazy var paper3Button = UIButton().then {
         $0.addTarget(self, action: #selector(didTapPaperButton(_:)), for: .touchUpInside)
         $0.backgroundColor = .paper3
         $0.tag = 3
@@ -70,7 +70,7 @@ class HomeViewController: UIViewController {
         $0.isHidden = true
     }
     
-    private lazy var paper4Button = UIButton().then {
+    lazy var paper4Button = UIButton().then {
         $0.addTarget(self, action: #selector(didTapPaperButton(_:)), for: .touchUpInside)
         $0.backgroundColor = .paper4
         $0.tag = 4
@@ -78,7 +78,7 @@ class HomeViewController: UIViewController {
         $0.isHidden = true
     }
     
-    private let whiteShadowView = UIImageView().then {
+    let whiteShadowView = UIImageView().then {
         $0.image = UIImage(named: "bgElementsStroke")
         $0.isHidden = true
     }
@@ -148,120 +148,6 @@ class HomeViewController: UIViewController {
     }
     
     // MARK: - Methods
-    
-    func setConstraint() {
-        view.addSubviews([backgroundView, whiteShadowView, dateLabel, guideLabel, arrowImage, trashBinButton, paper1Button, paper2Button, paper3Button, paper4Button])
-        
-        backgroundView.snp.makeConstraints { make in
-            make.top.leading.trailing.bottom.equalToSuperview()
-        }
-        
-        dateLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(self.view.safeAreaLayoutGuide).offset(SizeConstants.screenRatio * 13.0)
-        }
-        
-        guideLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(dateLabel.snp.bottom).offset(40)
-        }
-        
-        arrowImage.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(guideLabel.snp.bottom).offset(37 * SizeConstants.screenRatio)
-            make.width.equalToSuperview().multipliedBy(90.0 / 375.0)
-            make.height.equalTo(arrowImage.snp.width)
-        }
-        
-        trashBinButton.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(arrowImage.snp.bottom).offset(59 * SizeConstants.screenRatio)
-            make.width.equalToSuperview().multipliedBy(187.0 / 375.0)
-            make.height.equalToSuperview().multipliedBy(0.312)
-        }
-        
-        let width = 55.7 / SizeConstants.screenWidth
-        let height = 84.5 / width
-        
-        paper1Button.snp.makeConstraints { make in
-            make.top.equalTo(guideLabel.snp.bottom).offset(90.39 * SizeConstants.screenRatio)
-            make.leading.equalToSuperview().inset(33 * SizeConstants.screenRatio)
-            make.width.equalTo(self.view).multipliedBy(width)
-            make.height.equalTo(84.5).multipliedBy(height)
-        }
-        
-        paper2Button.snp.makeConstraints { make in
-            make.top.equalTo(guideLabel.snp.bottom).offset(47.77 * SizeConstants.screenRatio)
-            make.leading.equalTo(paper1Button.snp.trailing).offset(24.81 * SizeConstants.screenRatio)
-            make.width.equalTo(self.view).multipliedBy(width)
-            make.height.equalTo(84.5).multipliedBy(height)
-        }
-        
-        paper3Button.snp.makeConstraints { make in
-            make.top.equalTo(paper2Button.snp.top)
-            make.trailing.equalTo(paper4Button.snp.leading).offset(-24.81 * SizeConstants.screenRatio)
-            make.width.equalTo(self.view).multipliedBy(width)
-            make.height.equalTo(84.5).multipliedBy(height)
-        }
-        
-        paper4Button.snp.makeConstraints { make in
-            make.top.equalTo(paper1Button.snp.top)
-            make.trailing.equalToSuperview().inset(33 * SizeConstants.screenRatio)
-            make.width.equalTo(self.view).multipliedBy(width)
-            make.height.equalTo(84.5).multipliedBy(height)
-        }
-        
-        whiteShadowView.snp.makeConstraints { make in
-            make.top.equalTo(guideLabel.snp.bottom).offset(25)
-            make.width.equalToSuperview()
-            make.height.equalTo(whiteShadowView.snp.width).multipliedBy(264.0 / 375.0)
-        }
-    }
-    
-    func setView() {
-        view.backgroundColor = UIColor.blue2Main
-        
-        navigationController?.isNavigationBarHidden = true
-        navigationController?.interactivePopGestureRecognizer?.delegate = nil
-    }
-    
-    func configureInitAnimate() {
-        whiteShadowView.alpha = 0
-        paper1Button.alpha = 0
-        paper2Button.alpha = 0
-        paper3Button.alpha = 0
-        paper4Button.alpha = 0
-
-        whiteShadowView.transform = CGAffineTransform(translationX: 0, y: 100)
-        paper1Button.transform = CGAffineTransform(translationX: 0, y: 100)
-        paper2Button.transform = CGAffineTransform(translationX: 0, y: 100)
-        paper3Button.transform = CGAffineTransform(translationX: 0, y: 100)
-        paper4Button.transform = CGAffineTransform(translationX: 0, y: 100)
-    }
-    
-    func configureAnimate() {
-        animator.addAnimations {
-            self.whiteShadowView.alpha = 1
-            self.paper1Button.alpha = 1
-            self.paper2Button.alpha = 1
-            self.paper3Button.alpha = 1
-            self.paper4Button.alpha = 1
-            
-            self.whiteShadowView.transform = .identity
-            self.paper1Button.layer.transform = CATransform3DMakeRotation(-CGFloat(Double.pi / 180 * 45), 0, 0, 1)
-            self.paper2Button.layer.transform = CATransform3DMakeRotation(-CGFloat(Double.pi / 180 * 15), 0, 0, 1)
-            self.paper3Button.layer.transform = CATransform3DMakeRotation(CGFloat(Double.pi / 180 * 15), 0, 0, 1)
-            self.paper4Button.layer.transform = CATransform3DMakeRotation(CGFloat(Double.pi / 180 * 45), 0, 0, 1)
-        }
-    }
-    
-    func makeLabelTransition() {
-        let transition = CATransition()
-        transition.duration = 0.5
-        transition.timingFunction = .init(name: .easeInEaseOut)
-        transition.type = .fade
-        guideLabel.layer.add(transition, forKey: CATransitionType.push.rawValue)
-    }
     
     // MARK: - Protocols
 }
